Rename "lib" to "laser_setup"
