+++ laser_setup/instruments.py
